Refactor RepositoryProvider to utilize custom hooks for file data management and improve remote file fetching logic. Removed redundant code and enhanced error handling for fetching icons and variables from the repository.
