+++ apps/figma-plugin/src/hooks/useAdapterWorker.ts
@@ -1,0 +1,307 @@
+import { useCallback } from 'react';
+import type { RecursicaConfiguration } from '@recursica/schemas';
+import type { BaseRepository, Project } from '../services/repository';
+
+interface AdapterFile {
+  path: string;
+  content: string;
+}
+
+interface WorkerMessage {
+  bundledJson: string;
+  srcPath: string;
+  project: unknown;
+  rootPath: string;
+  iconsJson: string;
+  overrides?: unknown;
+  iconsConfig?: unknown;
+}
+
+interface WorkerResponse {
+  recursicaTokens?: { path: string; content: string };
+  vanillaExtractThemes: {
+    availableThemes?: { path: string; content: string };
+    themeContract?: { path: string; content: string };
+    themesFileContent?: { path: string; content: string };
+    vanillaExtractThemes: Array<{ path: string; content: string }>;
+  };
+  mantineTheme: {
+    mantineTheme?: { path: string; content: string };
+    postCss?: { path: string; content: string };
+  };
+  uiKitObject?: { path: string; content: string };
+  recursicaObject?: { path: string; content: string };
+  colorsType?: { path: string; content: string };
+  iconsObject?: {
+    iconExports: { path: string; content: string };
+    iconResourceMap: { path: string; content: string };
+    exportedIcons: Array<{ path: string; content: string }>;
+  };
+  prettierignore?: { path: string; content: string };
+}
+
+export function useAdapterWorker() {
+  const runAdapter = useCallback(
+    async (
+      repositoryInstance: BaseRepository,
+      selectedProject: Project,
+      targetBranch: string,
+      config: RecursicaConfiguration,
+      fileLoadingData: {
+        localIconsJson: string | null;
+        localBundledJson: string | null;
+        remoteIconsJson: string | null;
+        remoteBundledJson: string | null;
+      },
+      onStatusUpdate: (status: 'loading' | 'done' | 'error') => void
+    ): Promise<AdapterFile[]> => {
+      if (!repositoryInstance || !selectedProject) {
+        throw new Error('Failed to run adapter');
+      }
+
+      let rootPath = '';
+      if (typeof config.project === 'object' && config.project.root) {
+        rootPath = config.project.root;
+      }
+
+      let adapterPath = 'adapter.js';
+      if (rootPath) {
+        adapterPath = rootPath + '/' + adapterPath;
+      }
+      if (typeof config.project === 'object' && config.project.adapter) {
+        adapterPath = config.project.adapter;
+      }
+
+      let adapterFile = null;
+      try {
+        adapterFile = await repositoryInstance.getSingleFile(
+          selectedProject,
+          adapterPath,
+          targetBranch
+        );
+        onStatusUpdate('loading');
+      } catch (error) {
+        onStatusUpdate('error');
+        if (error instanceof Error && error.message.includes('404')) {
+          console.error('Adapter file not found');
+          return [];
+        }
+        throw error;
+      }
+
+      // Use local data if available, otherwise use remote data
+      const iconsJson = fileLoadingData.localIconsJson || fileLoadingData.remoteIconsJson || '{}';
+      const bundledJson =
+        fileLoadingData.localBundledJson || fileLoadingData.remoteBundledJson || '{}';
+
+      return new Promise<AdapterFile[]>((resolve, reject) => {
+        let worker: Worker;
+
+        try {
+          // Create worker with error handling
+          worker = new Worker(
+            URL.createObjectURL(new Blob([adapterFile], { type: 'text/javascript' }))
+          );
+        } catch (error) {
+          console.error('❌ Failed to create worker:', error);
+          onStatusUpdate('error');
+          reject(
+            new Error(
+              `Failed to create worker: ${error instanceof Error ? error.message : 'Unknown error'}`
+            )
+          );
+          return;
+        }
+
+        // Set a timeout for the worker (5 minutes)
+        const timeoutId = setTimeout(
+          () => {
+            console.error('❌ Worker execution timed out after 5 minutes');
+            worker.terminate();
+            onStatusUpdate('error');
+            reject(new Error('Worker execution timed out after 5 minutes'));
+          },
+          5 * 60 * 1000
+        );
+
+        try {
+          const message: WorkerMessage = {
+            bundledJson,
+            srcPath: rootPath + '/src',
+            project: config.project,
+            rootPath,
+            iconsJson,
+            overrides: config.overrides,
+            iconsConfig: config.icons,
+          };
+          worker.postMessage(message);
+        } catch (error) {
+          clearTimeout(timeoutId);
+          worker.terminate();
+          console.error('❌ Failed to post message to worker:', error);
+          onStatusUpdate('error');
+          reject(
+            new Error(
+              `Failed to post message to worker: ${error instanceof Error ? error.message : 'Unknown error'}`
+            )
+          );
+          return;
+        }
+
+        // Listener for messages coming FROM the worker
+        worker.onmessage = (event) => {
+          clearTimeout(timeoutId); // Clear timeout on successful response
+          worker.terminate(); // Clean up worker
+          console.log('✅ Response received from worker:', event.data);
+
+          onStatusUpdate('done');
+          const response: WorkerResponse = event.data;
+          const {
+            recursicaTokens,
+            vanillaExtractThemes,
+            mantineTheme,
+            uiKitObject,
+            recursicaObject,
+            colorsType,
+            iconsObject,
+            prettierignore,
+          } = response;
+          const newAdapterFiles: AdapterFile[] = [];
+
+          if (recursicaTokens) {
+            newAdapterFiles.push({
+              path: recursicaTokens.path,
+              content: recursicaTokens.content,
+            });
+          }
+
+          if (prettierignore) {
+            newAdapterFiles.push({
+              path: prettierignore.path,
+              content: prettierignore.content,
+            });
+          }
+
+          const {
+            availableThemes,
+            themeContract,
+            themesFileContent,
+            vanillaExtractThemes: subThemes,
+          } = vanillaExtractThemes;
+          if (availableThemes) {
+            newAdapterFiles.push({
+              path: availableThemes.path,
+              content: availableThemes.content,
+            });
+          }
+          if (themeContract) {
+            newAdapterFiles.push({
+              path: themeContract.path,
+              content: themeContract.content,
+            });
+          }
+          if (themesFileContent) {
+            newAdapterFiles.push({
+              path: themesFileContent.path,
+              content: themesFileContent.content,
+            });
+          }
+          for (const theme of subThemes) {
+            newAdapterFiles.push({
+              path: theme.path,
+              content: theme.content,
+            });
+          }
+
+          if (mantineTheme.mantineTheme) {
+            newAdapterFiles.push({
+              path: mantineTheme.mantineTheme.path,
+              content: mantineTheme.mantineTheme.content,
+            });
+          }
+          if (mantineTheme.postCss) {
+            newAdapterFiles.push({
+              path: mantineTheme.postCss.path,
+              content: mantineTheme.postCss.content,
+            });
+          }
+
+          if (uiKitObject) {
+            newAdapterFiles.push({
+              path: uiKitObject.path,
+              content: uiKitObject.content,
+            });
+          }
+
+          if (recursicaObject) {
+            newAdapterFiles.push({
+              path: recursicaObject.path,
+              content: recursicaObject.content,
+            });
+          }
+
+          if (colorsType) {
+            newAdapterFiles.push({
+              path: colorsType.path,
+              content: colorsType.content,
+            });
+          }
+
+          if (iconsObject) {
+            newAdapterFiles.push({
+              path: iconsObject.iconExports.path,
+              content: iconsObject.iconExports.content,
+            });
+            newAdapterFiles.push({
+              path: iconsObject.iconResourceMap.path,
+              content: iconsObject.iconResourceMap.content,
+            });
+            for (const icon of iconsObject.exportedIcons) {
+              newAdapterFiles.push({
+                path: icon.path,
+                content: icon.content,
+              });
+            }
+          }
+          resolve(newAdapterFiles);
+        };
+
+        // Listener for any errors that might occur inside the worker
+        worker.onerror = (error) => {
+          clearTimeout(timeoutId); // Clear timeout on error
+          worker.terminate(); // Clean up worker
+          console.error('❌ Error in worker:', error);
+
+          // Update UI status immediately
+          onStatusUpdate('error');
+
+          // Create a more detailed error message
+          const errorMessage = error.message || 'Unknown worker error';
+          const errorDetails = error.filename
+            ? `File: ${error.filename}, Line: ${error.lineno}`
+            : '';
+          const fullError = `Worker execution failed: ${errorMessage}${errorDetails ? ` (${errorDetails})` : ''}`;
+
+          reject(new Error(fullError));
+        };
+
+        // Handle worker message errors
+        worker.onmessageerror = (error) => {
+          clearTimeout(timeoutId); // Clear timeout on message error
+          worker.terminate(); // Clean up worker
+          console.error('❌ Message error in worker:', error);
+
+          // Update UI status immediately
+          onStatusUpdate('error');
+
+          reject(new Error(`Worker message error: Unable to process worker message`));
+        };
+      });
+    },
+    []
+  );
+
+  return {
+    runAdapter,
+  };
+}
